requirements.txt
yfinance==0.2.32
pandas==3.0.1
requests==2.33.0
plotly==6.6.0
rich==14.3.3
google-generativeai==0.8.6
ta==0.11.0
